+++ ccramic/callbacks/roi_level_callbacks.py
@@ -37,11 +37,16 @@ def init_roi_level_callbacks(dash_app, tmpdirname, authentic_id):
                        State('mask-dict', 'data'),
                        State('data-collection', 'options'),
                        State('query-cell-id-lists', 'data'),
+                       State('bool-apply-global-filter', 'value'),
+                       State('global-filter-type', 'value'),
+                       State("global-kernel-val-filter", 'value'),
+                       State("global-sigma-val-filter", 'value'),
                        prevent_initial_call=True)
     def generate_roi_images_from_query(currently_selected, data_selection, blend_colour_dict,
                                     session_config, execute_query, num_queries, rois_exclude, load_additional,
                                     existing_gallery, execute_quant_query, query_from_quantification, mask_dict,
-                                    dataset_options, query_cell_id_lists):
+                                    dataset_options, query_cell_id_lists, global_apply_filter,
+                                    global_filter_type, global_filter_val, global_filter_sigma,):
         """
         Generate the dynamic gallery of ROI queries from the query selection
         Can be activated using either the original button for a fresh query, or the button to load additional ROIs
@@ -66,7 +71,7 @@ def init_roi_level_callbacks(dash_app, tmpdirname, authentic_id):
                 row_children = []
             images = generate_multi_roi_images_from_query(session_config, blend_colour_dict,
                     currently_selected, int(num_queries), rois_exclude, rois_decided, mask_dict, dataset_options,
-                                                          query_cell_id_lists)
+                    query_cell_id_lists, global_apply_filter, global_filter_type, global_filter_val, global_filter_sigma)
             new_row_children, roi_list = generate_roi_query_gallery_children(images)
             # if the query is being extended, append to the existing gallery for exclusion. Otherwise, start fresh
             if ctx.triggered_id == "dataset-query-additional-load":